Updated types of fields of ProductEntity

+++ src/main/java/com/epam/druzhinin/entity/ProductEntity.java
@@ -3,8 +3,8 @@ package com.epam.druzhinin.entity;
 import lombok.Data;
 import lombok.experimental.Accessors;
 
-
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 @Data
@@ -24,11 +24,11 @@ public class ProductEntity {
     @Column(name = "category")
     private String category;
 
-    @Column(name="amount")
+    @Column(name = "amount")
     private Integer amount;
 
-    @Column(name="price")
-    private Integer price;
+    @Column(name = "price")
+    private BigDecimal price;
 
     @Column(name = "date", nullable = false)
     private ZonedDateTime date;
